+++ archives/HISTORY.md
@@ -1,5 +1,11 @@
 # HISTORY
 
+## 2018-06-14
+
+* Fix bug in graph
+* Change legend click behavior
+* Fix Layer group
+
 ## 2017-12-06
 
 * upgrade GeoServer to 2.13 and Marlin 0.8.2